requirements.txt
numpy==1.26.4 
Pillow==10.2.0